improve grapesjs with split code
